feat!: report the shared backend in Stats, and settle the counter types
Two changes to the same struct, so one commit.

**Nothing in Stats said whether the shared backend was being reached.** No
count of calls, refusals, failures or breaker short-circuits — so an operator
whose Redis was down saw a snapshot that looked healthy while every replica
quietly fell back to enforcing the rate per process, which is exactly the
failure mode QuotaFallbackLocal is designed to produce and exactly the one you
want to know about. QuotaTakes, QuotaRefused and QuotaErrors close that.
QuotaErrors is the number to alert on, and it counts short-circuited calls as
well as failed ones: from the caller's side they are the same event, and the
breaker being open is itself evidence the backend is down.

**Requests, Throttled, Errors and Evictions become int64.** Field types cannot
change after v1, so this is the last chance. Metrics code diffs consecutive
snapshots and unsigned subtraction wraps silently on the reset a restarted
Limiter produces — a negative delta is a visible bug where
18446744073709551615 is a mystery. Users was already int64, so the struct was
mixing signedness for no reason.

**Wait becomes WaitTotal.** It is a running sum; "Wait" alone reads as a
current or average value, and nothing in the name recovers the difference.

Stats.Throttled's doc now also records where it under-counts: with a
WaitingSharedQuota the backend owns the wait, so pace cannot tell in advance
whether a caller will be parked and reports nothing. Better to write the gap
down than to let someone find it by diffing two counters that should agree.

+++ limiter.go
@@ -732,7 +732,7 @@ func (l *Limiter) dropUsers() {
 	if !notify {
 		// Nobody is listening, so the per-user list was never built; count the
 		// drop in one go instead.
-		l.stats.evictions.Add(uint64(max(0, dropped)))
+		l.stats.evictions.Add(max(0, dropped))
 	}
 }
